Moved CounterHandler class to a different file
